+++ packages/rich-text/src/plugins/shared/FetchingWrappedEntryCard.tsx
@@ -2,14 +2,18 @@ import * as React from 'react';
 
 import { FieldAppSDK } from '@contentful/app-sdk';
 import { ScheduledAction, Entry } from '@contentful/app-sdk';
-import { EntryCard } from '@contentful/f36-components';
+import { DragHandle, EntryCard } from '@contentful/f36-components';
 import {
   useEntity,
   MissingEntityCard,
   WrappedEntryCard,
   useEntityLoader,
 } from '@contentful/field-editor-reference';
-import { LocalePublishStatusMap, useLocalePublishStatus } from '@contentful/field-editor-shared';
+import {
+  LocalePublishStatusMap,
+  useLocalePublishStatus,
+  useActiveLocales,
+} from '@contentful/field-editor-shared';
 import areEqual from 'fast-deep-equal';
 
 interface InternalEntryCard {
@@ -30,6 +34,7 @@ const InternalEntryCard = React.memo((props: InternalEntryCard) => {
   const contentType = sdk.space
     .getCachedContentTypes()
     .find((contentType) => contentType.sys.id === entry.sys.contentType.sys.id);
+  const activeLocales = useActiveLocales(props.sdk);
 
   return (
     <WrappedEntryCard
@@ -47,7 +52,12 @@ const InternalEntryCard = React.memo((props: InternalEntryCard) => {
       isClickable={false}
       useLocalizedEntityStatus={sdk.parameters.instance.useLocalizedEntityStatus}
       localesStatusMap={props.localesStatusMap}
-      activeLocales={props.sdk.parameters.instance.activeLocales}
+      activeLocales={activeLocales}
+      renderDragHandle={
+        !props.isDisabled
+          ? (dragHandleProps) => <DragHandle label="drag embedded entry" {...dragHandleProps} />
+          : undefined
+      }
     />
   );
 }, areEqual);