+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.16](https://github.com/contentful/field-editors/compare/@contentful/field-editor-radio@1.5.15...@contentful/field-editor-radio@1.5.16) (2025-03-10)
+
+**Note:** Version bump only for package @contentful/field-editor-radio
+
 ## [1.5.15](https://github.com/contentful/field-editors/compare/@contentful/field-editor-radio@1.5.14...@contentful/field-editor-radio@1.5.15) (2025-02-04)
 
 **Note:** Version bump only for package @contentful/field-editor-radio