+++ packages/validation-errors/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-validation-errors@1.4.0...@contentful/field-editor-validation-errors@1.4.1) (2025-03-10)
+
+**Note:** Version bump only for package @contentful/field-editor-validation-errors
+
 # [1.4.0](https://github.com/contentful/field-editors/compare/@contentful/field-editor-validation-errors@1.3.33...@contentful/field-editor-validation-errors@1.4.0) (2025-02-19)
 
 ### Features